+++ packages/flutter_bloc/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 7.0.0-dev.1
+
+- **BREAKING**: remove dependency on `package:provider`
+- feat: add optional `listen` to `BlocProvider`
+- feat: add `ListenProviderExtension`
+- docs: improve inline documentation for `BlocProvider` and `RepositoryProvider`
+
 # 6.0.6
 
 - docs: improve inline documentation for `buildWhen`